Fix clang-tidy 23 findings
clang-tidy 23 added new checks and renamed some existing ones, failing
the tidy build with ~188 unique findings across 22 check names.

Disable the new noisy checks in .clang-tidy (per the policy at the top
of the file), including the new canonical names of already-disabled
checks (`bugprone-signed-bitwise` for `hicpp-signed-bitwise`,
`portability-no-assembler` for `hicpp-no-assembler`) and the new
aliases of `google-explicit-constructor`, whose `NOLINT`s in the code
reference the google name.

Fix the small mechanical findings:
- `modernize-redundant-void-arg`: `std::function<void(void)>` and
  similar in 7 places.
- `readability-redundant-casting`: remove 11 redundant casts.
- `bugprone-unsafe-to-allow-exceptions`: `NOLINT` the deliberately
  throwing `~SimpleFaultInjection`.

Also disable the `lifetime-safety-dangling-field` compiler diagnostic
for clang >= 23: it does not understand a field that is deliberately
cleared (via `SCOPE_EXIT`) before the captured local dies.

All touched translation units compile cleanly in a local debug build.

https://s3.amazonaws.com/clickhouse-test-reports/praktika.html?PR=117346&sha=ed8367eb17b594dbb12b741e1354473f53f82c11&name_0=PR&name_1=Build%20%28arm_tidy%29
https://github.com/ClickHouse/ClickHouse/pull/117346

+++ src/Client/BuzzHouse/Generator/SQLTable.cpp
@@ -1409,26 +1409,26 @@ void StatementGenerator::generateEngineDetails(
         else if (b.isBufferEngine())
         {
             /// num_layers
-            te->add_params()->set_num(static_cast<uint32_t>(rg.nextLargeNumber() % 101));
+            te->add_params()->set_num(rg.nextLargeNumber() % 101);
             /// min_time, max_time, min_rows, max_rows, min_bytes, max_bytes
             for (int i = 0; i < 6; i++)
             {
-                te->add_params()->set_num(static_cast<uint32_t>(rg.nextLargeNumber() % 1001));
+                te->add_params()->set_num(rg.nextLargeNumber() % 1001);
             }
             if (rg.nextSmallNumber() < 7)
             {
                 /// flush_time
-                te->add_params()->set_num(static_cast<uint32_t>(rg.nextLargeNumber() % 61));
+                te->add_params()->set_num(rg.nextLargeNumber() % 61);
             }
             if (rg.nextSmallNumber() < 7)
             {
                 /// flush_rows
-                te->add_params()->set_num(static_cast<uint32_t>(rg.nextLargeNumber() % 1001));
+                te->add_params()->set_num(rg.nextLargeNumber() % 1001);
             }
             if (rg.nextSmallNumber() < 7)
             {
                 /// flush_bytes
-                te->add_params()->set_num(static_cast<uint32_t>(rg.nextLargeNumber() % 1001));
+                te->add_params()->set_num(rg.nextLargeNumber() % 1001);
             }
         }
     }
