+++ src/Common/Increment.h
@@ -48,7 +48,7 @@ public:
       */
     UInt64 getBunch(UInt64 count, bool create_if_need = false)
     {
-        return static_cast<UInt64>(counter.add(static_cast<Int64>(count), create_if_need) - count + 1);
+        return counter.add(static_cast<Int64>(count), create_if_need) - count + 1;
     }
 
     /// Change the path to the file.
